+++ tango/server/index.js
@@ -6,13 +6,38 @@ fetch("/api/runlist", {
 })
   .then((res) => res.json())
   .then((out) => {
-    out.forEach((val) => {
-      const div = document.createElement("div");
-      const ahref = document.createElement("a");
-      ahref.setAttribute("href", `/run/${val}`);
-      ahref.innerHTML = val;
-      div.appendChild(ahref);
-      document.getElementById("content").appendChild(div);
+    const table = document.createElement("table");
+    document.getElementById("content").appendChild(table);
+    table.className='table';
+    const trHead = document.createElement("tr");
+    table.appendChild(trHead);
+    trHead.className='table-header';
+    const thName = document.createElement("th");
+    trHead.appendChild(thName);
+    thName.className='header__item';
+    thName.innerHTML = "Run Name";
+    const thStartDate = document.createElement("th");
+    trHead.appendChild(thStartDate);
+    thStartDate.className='header__item';
+    thStartDate.innerHTML = "Start Date";
+    out
+      .map(v => { return { name:v.name, startDate:dayjs(v.start_date) } })
+      .sort((a, b) => b.startDate-a.startDate)
+      .forEach((val) => {
+        const tr = document.createElement("tr");
+        table.appendChild(tr);
+        tr.className='table-row';
+        const tdName = document.createElement("td");
+        tr.appendChild(tdName);
+        tdName.className='table-data';
+        const ahref = document.createElement("a");
+        tdName.appendChild(ahref);
+        ahref.setAttribute("href", `/run/${val.name}`);
+        ahref.innerHTML = val.name;
+        const tdStartDate = document.createElement("td");
+        tr.appendChild(tdStartDate);
+        tdStartDate.className='table-data';
+        tdStartDate.innerHTML = val.startDate.format("YYYY-MM-DD h:mm:ss A");
     });
   })
   .catch((err) => {